LPS-174758 LPS-200418 Refactor assetViewURL so the code can be reused arround components
