+++ modules/apps/portal-search/portal-search-api/src/main/java/com/liferay/portal/search/asset/AssetURLViewProvider.java
@@ -1,0 +1,27 @@
+/**
+ * SPDX-FileCopyrightText: (c) 2023 Liferay, Inc. https://liferay.com
+ * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
+ */
+
+package com.liferay.portal.search.asset;
+
+import aQute.bnd.annotation.ProviderType;
+
+import com.liferay.asset.kernel.model.AssetRenderer;
+import com.liferay.asset.kernel.model.AssetRendererFactory;
+import com.liferay.portal.kernel.portlet.LiferayPortletRequest;
+import com.liferay.portal.kernel.portlet.LiferayPortletResponse;
+
+/**
+ * @author Gustavo Lima
+ */
+@ProviderType
+public interface AssetURLViewProvider {
+
+	public String getAssetURLView(
+		AssetRenderer<?> assetRenderer,
+		AssetRendererFactory<?> assetRendererFactory, String className,
+		long classPK, LiferayPortletRequest liferayPortletRequest,
+		LiferayPortletResponse liferayPortletResponse);
+
+}
