+++ modules/apps/portal-search/portal-search/src/main/java/com/liferay/portal/search/internal/asset/AssetURLViewProviderImpl.java
@@ -3,12 +3,12 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-package com.liferay.portal.search.internal.util;
+package com.liferay.portal.search.internal.asset;
 
 import com.liferay.asset.kernel.model.AssetEntry;
 import com.liferay.asset.kernel.model.AssetRenderer;
 import com.liferay.asset.kernel.model.AssetRendererFactory;
-import com.liferay.asset.kernel.service.AssetEntryLocalServiceUtil;
+import com.liferay.asset.kernel.service.AssetEntryLocalService;
 import com.liferay.petra.string.StringBundler;
 import com.liferay.petra.string.StringPool;
 import com.liferay.portal.kernel.log.Log;
@@ -18,9 +18,10 @@ import com.liferay.portal.kernel.portlet.LiferayPortletResponse;
 import com.liferay.portal.kernel.portlet.url.builder.PortletURLBuilder;
 import com.liferay.portal.kernel.theme.ThemeDisplay;
 import com.liferay.portal.kernel.util.HttpComponentsUtil;
-import com.liferay.portal.kernel.util.PortalUtil;
+import com.liferay.portal.kernel.util.Portal;
 import com.liferay.portal.kernel.util.Validator;
 import com.liferay.portal.kernel.util.WebKeys;
+import com.liferay.portal.search.asset.AssetURLViewProvider;
 import com.liferay.portal.search.web.constants.SearchResultsPortletKeys;
 
 import javax.portlet.MutableRenderParameters;
@@ -29,12 +30,17 @@ import javax.portlet.PortletRequest;
 import javax.portlet.PortletURL;
 import javax.portlet.WindowState;
 
+import org.osgi.service.component.annotations.Component;
+import org.osgi.service.component.annotations.Reference;
+
 /**
  * @author Gustavo Lima
  */
-public class AssetURLUtil {
+@Component(service = AssetURLViewProvider.class)
+public class AssetURLViewProviderImpl implements AssetURLViewProvider {
 
-	public static String getAssetURLView(
+	@Override
+	public String getAssetURLView(
 		AssetRenderer<?> assetRenderer,
 		AssetRendererFactory<?> assetRendererFactory, String className,
 		long classPK, LiferayPortletRequest liferayPortletRequest,
@@ -47,7 +53,7 @@ public class AssetURLUtil {
 					SearchResultsPortletKeys.SEARCH_RESULTS,
 					PortletRequest.RENDER_PHASE
 				).setRedirect(
-					PortalUtil.getCurrentURL(liferayPortletRequest)
+					_portal.getCurrentURL(liferayPortletRequest)
 				).setPortletMode(
 					PortletMode.VIEW
 				).setWindowState(
@@ -59,7 +65,7 @@ public class AssetURLUtil {
 
 			mutableRenderParameters.setValue("mvcPath", "/view_content.jsp");
 
-			AssetEntry assetEntry = AssetEntryLocalServiceUtil.getEntry(
+			AssetEntry assetEntry = _assetEntryLocalService.getEntry(
 				className, classPK);
 
 			mutableRenderParameters.setValue(
@@ -90,7 +96,7 @@ public class AssetURLUtil {
 		catch (Exception exception) {
 			_log.error(
 				StringBundler.concat(
-					"Unable to get view URL for class ", className,
+					"Unable to get asset view URL for class ", className,
 					" with primary key ", classPK),
 				exception);
 		}
@@ -98,6 +104,13 @@ public class AssetURLUtil {
 		return StringPool.BLANK;
 	}
 
-	private static final Log _log = LogFactoryUtil.getLog(AssetURLUtil.class);
+	private static final Log _log = LogFactoryUtil.getLog(
+		AssetURLViewProviderImpl.class);
+
+	@Reference
+	private AssetEntryLocalService _assetEntryLocalService;
+
+	@Reference
+	private Portal _portal;
 
 }
