+++ modules/apps/portal-search/portal-search/src/main/java/com/liferay/portal/search/internal/suggestions/spi/asah/user/activity/RecentAssetsUserActivityAsahSuggestionsContributor.java
@@ -14,7 +14,7 @@ import com.liferay.portal.kernel.portlet.LiferayPortletRequest;
 import com.liferay.portal.kernel.portlet.LiferayPortletResponse;
 import com.liferay.portal.kernel.search.SearchContext;
 import com.liferay.portal.kernel.util.HashMapBuilder;
-import com.liferay.portal.search.internal.util.AssetURLUtil;
+import com.liferay.portal.search.asset.AssetURLViewProvider;
 import com.liferay.portal.search.rest.dto.v1_0.SuggestionsContributorConfiguration;
 import com.liferay.portal.search.spi.suggestions.SuggestionsContributor;
 import com.liferay.portal.search.suggestions.SuggestionsContributorResults;
@@ -23,6 +23,7 @@ import com.liferay.portal.search.suggestions.spi.constants.AsahSuggestionsConsta
 import java.util.HashMap;
 
 import org.osgi.service.component.annotations.Component;
+import org.osgi.service.component.annotations.Reference;
 
 /**
  * @author Gustavo Lima
@@ -74,7 +75,7 @@ public class RecentAssetsUserActivityAsahSuggestionsContributor
 
 				long classPK = itemJSONObject.getLong("assetId");
 
-				return AssetURLUtil.getAssetURLView(
+				return _assetURLViewProvider.getAssetURLView(
 					assetRendererFactory.getAssetRenderer(classPK),
 					assetRendererFactory, className, classPK,
 					_liferayPortletRequest, _liferayPortletResponse);
@@ -105,6 +106,9 @@ public class RecentAssetsUserActivityAsahSuggestionsContributor
 		).put(
 			"web-content", "com.liferay.journal.model.JournalArticle"
 		).build();
+
+	@Reference
+	private AssetURLViewProvider _assetURLViewProvider;
 
 	private LiferayPortletRequest _liferayPortletRequest;
 	private LiferayPortletResponse _liferayPortletResponse;
